chore: move code for creaing pod specs to a separate file
